Refactor test helpers and add VB tests
